backend/requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.7
django-appconf==1.1.0
django-cors-headers==4.9.0
django-filter==25.2
django-imagekit==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
graphviz==0.21
idna==3.11
kombu==5.5.4
packaging==25.0
pilkit==3.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydot==4.0.1
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==6.4.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
